fix(filters): dedupe globToLikePattern, pin test against live impl
TestGlobToLikePattern was pinning a byte-identical but unreachable copy of
globToLikePattern in internal/storage/issueops/filters.go — the production
caller is internal/storage/sqlbuild/filter.go:325 via BuildIssueFilterClauses.
An escaping regression in the sqlbuild copy (%, _, | handling) could have
shipped unseen since only the dead copy was tested.

Delete the issueops copy (and its now-unused "strings" import) and move
TestGlobToLikePattern into internal/storage/sqlbuild, in-package, so it
exercises the function BuildIssueFilterClauses actually calls.

Flagged by cross-vendor paired review (claude-opus + codex-gpt-5.6-sol) on
PR #3971.

Agent-Signature: claude-sonnet-medium on behalf of matt wilkie
