+++ internal/storage/issueops/filters.go
@@ -1,8 +1,6 @@
 package issueops
 
 import (
-	"strings"
-
 	"github.com/steveyegge/beads/internal/storage/sqlbuild"
 	"github.com/steveyegge/beads/internal/types"
 )
@@ -17,29 +15,6 @@ var (
 	IssuesFilterTables = sqlbuild.IssuesFilterTables
 	WispsFilterTables  = sqlbuild.WispsFilterTables
 )
-
-// globToLikePattern converts a shell-style glob (* and ?) to a SQL LIKE
-// pattern. Literal % and _ in the input — and the '|' escape char itself —
-// are escaped so they don't act as LIKE wildcards. The resulting SQL must
-// use ESCAPE '|'.
-func globToLikePattern(pattern string) string {
-	var b strings.Builder
-	b.Grow(len(pattern))
-	for _, c := range pattern {
-		switch c {
-		case '%', '_', '|':
-			b.WriteByte('|')
-			b.WriteRune(c)
-		case '*':
-			b.WriteByte('%')
-		case '?':
-			b.WriteByte('_')
-		default:
-			b.WriteRune(c)
-		}
-	}
-	return b.String()
-}
 
 // BuildIssueFilterClauses builds WHERE clause fragments and args from a query
 // string and IssueFilter. The tables parameter controls which table names are
