+++ internal/storage/issueops/filters_test.go
@@ -352,35 +352,6 @@ func TestBuildIssueFilterClauses_LabelRegex(t *testing.T) {
 	}
 }
 
-func TestGlobToLikePattern(t *testing.T) {
-	t.Parallel()
-
-	cases := []struct {
-		name string
-		in   string
-		want string
-	}{
-		{name: "trailing star", in: "tech-*", want: "tech-%"},
-		{name: "surrounding stars", in: "*foo*", want: "%foo%"},
-		{name: "question mark", in: "v?", want: "v_"},
-		{name: "literal percent", in: "5%", want: "5|%"},
-		{name: "literal underscore", in: "snake_case", want: "snake|_case"},
-		{name: "literal pipe", in: "a|b", want: "a||b"},
-		{name: "no metachars", in: "needs-pm", want: "needs-pm"},
-		{name: "empty", in: "", want: ""},
-	}
-	for _, tc := range cases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			got := globToLikePattern(tc.in)
-			if got != tc.want {
-				t.Errorf("globToLikePattern(%q) = %q, want %q", tc.in, got, tc.want)
-			}
-		})
-	}
-}
-
 func TestBuildIssueFilterClauses_DateFilters(t *testing.T) {
 	t.Parallel()
